Steves q debug logger alpha (#89)
* Guten Morgen und bis gleich (#75)

Fixed Issue #73

* Update btw17.config (#78)

* Update btw17.config

Issue #68

* Update btw17.config

Issue #68

* Add files via upload (#79)

* Add files via upload

the light version is the simplest version of design ideas but also the blueprint that shows the interaction of single parts (color; outline; wide range; shape; curved edge; background)
these will be uploaded collected tomorrow in different versions

* Add files via upload

* update settings.py

* Add files via upload

WARNIG: These files are only intended for early tests. There is a major problem with Strings that have to be fixed before merging to the Master branch. Nevertheless is QDebug Logger stable for testing purposes and is able to wirte to the console and a text file.

* updated logger